+++ src/core/se/jbee/inject/event/AsyncEventProcessor.java
@@ -1,0 +1,17 @@
+package se.jbee.inject.event;
+
+public class AsyncEventProcessor implements EventProcessor {
+
+	@Override
+	public <T> void register(Class<T> event, T listener) {
+		// TODO Auto-generated method stub
+		
+	}
+
+	@Override
+	public <T> T proxy(Class<T> event) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+}