+++ src/test/se/jbee/inject/bind/TestSupplierBinds.java
@@ -21,7 +21,7 @@ public class TestSupplierBinds {
 		}
 
 		@Override
-		public String supply( Dependency<? super String> dependency, Injector injector ) {
+		public String supply( Dependency<? super String> dep, Injector injector ) {
 			return "foobar";
 		}